+++ app/Collection.php
@@ -2,6 +2,7 @@
 
 namespace App;
 
+use App\Concerns\HasHeaderImage;
 use Astrotomic\Translatable\Contracts\Translatable as TranslatableContract;
 use Astrotomic\Translatable\Translatable;
 use Carbon\Carbon;
@@ -9,13 +10,17 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Query\JoinClause;
 use Illuminate\Support\Facades\URL;
-use Intervention\Image\ImageManagerStatic;
 
 class Collection extends Model implements TranslatableContract
 {
     use Translatable;
 
-    const ARTWORKS_DIR = '/images/kolekcie/';
+    use HasHeaderImage;
+
+    function getArtworksDirAttribute()
+    {
+        return '/images/kolekcie/';
+    }
 
     public $translatedAttributes = ['name','type', 'text'];
 
@@ -28,6 +33,12 @@ class Collection extends Model implements TranslatableContract
         'published_at' => 'sortable.published_at',
         'updated_at' => 'sortable.updated_at',
         'name'       => 'sortable.title',
+    );
+
+    protected $dates = array(
+        'created_at',
+        'updated_at',
+        'published_at'
     );
 
     public function items()
@@ -57,40 +68,6 @@ class Collection extends Model implements TranslatableContract
         $string = $striped_string;
         $string = substr($string, 0, $length);
         return ($striped_string > $string) ? substr($string, 0, strrpos($string, ' ')) . " ..." : $string;
-    }
-
-    public function hasHeaderImage()
-    {
-        return file_exists(self::getHeaderImageForId($this->id, true));
-    }
-
-    public function getHeaderImage($full = false)
-    {
-        return self::getHeaderImageForId($this->id, $full);
-    }
-
-    public static function getHeaderImageForId($id, $full = false)
-    {
-        $relative_path = self::ARTWORKS_DIR . $id . '.jpg';
-        $path = ($full) ? public_path() . $relative_path : $relative_path;
-        return $path;
-    }
-
-    public function getResizedImage($resize)
-    {
-        $file =  $this->id;
-        $full_path = public_path() .  self::ARTWORKS_DIR;
-        if (!file_exists($full_path . "$file.$resize.jpg")) {
-            try {
-                $img = \Image::make($this->getHeaderImage(true))->fit($resize)->sharpen(7);
-            } catch (\Exception $e) {
-                $img = \Image::make(public_path('images/no-image/no-image.jpg'))->fit($resize)->sharpen(7);
-            }
-
-            $img->save($full_path . "$file.$resize.jpg");
-        }
-        $result_path = self::ARTWORKS_DIR .  "$file.$resize.jpg";
-        return $result_path;
     }
 
     public function getContentImages()
@@ -147,5 +124,9 @@ class Collection extends Model implements TranslatableContract
                 $query->whereNotIn("$alias.collection_id", $withTranslation);
             }
         });
+    }
+
+    public function getReadingTimeAttribute(){
+        return getEstimatedReadingTime($this->text, \App::getLocale() );
     }
 }