+++ resources/lang/cs/articles.php
@@ -1,0 +1,14 @@
+<?php
+
+return array(
+    'count' => 'článek|články|článků',
+    'no_results' => 'momentálně žádné články',
+    'filter' => [
+        'author' => 'autor',
+        'category' => 'kategorie',
+        'sort_by' => [
+            'date_desc' => 'data',
+            'date_asc' => 'data od nejstarších',
+        ]
+    ]
+);