+++ resources/lang/en/articles.php
@@ -1,0 +1,15 @@
+<?php
+
+return array(
+    'title' => 'articles',
+    'count' => 'article|articles',
+    'no_results' => 'currently no articles',
+    'filter' => [
+        'author' => 'author',
+        'category' => 'category',
+        'sort_by' => [
+            'date_desc' => 'date',
+            'date_asc' => 'date, oldest first',
+        ]
+    ]
+);