+++ resources/views/components/searchbar.blade.php
@@ -12,7 +12,7 @@
         'autocomplete'=>'off',
         'data-artists-title'=> utrans('authority.authors'),
         'data-artworks-title'=> utrans('katalog.title'),
-        'data-articles-title'=> utrans('clanky.title'),
+        'data-articles-title'=> utrans('articles.title'),
         'data-collections-title'=> utrans('kolekcie.title'),
     ]) !!}
     {!!  Form::submit('submit'); !!}