+++ resources/views/frontend/authors/index.blade.php
@@ -106,9 +106,6 @@
 @section('javascript')
 
     {!! Html::script('js/bootstrap-slider.min.js') !!}
-    {!! Html::script('js/selectize.min.js') !!}
-
-    {!! Html::script('js/slick.js') !!}
     {!! Html::script('js/components/artwork_carousel.js') !!}
     {!! Html::script('js/jquery.dropdown-select.js') !!}