+++ resources/views/kolekcia.blade.php
@@ -2,16 +2,16 @@
 
 @section('og')
 @if (!$collection->publish)
-    <meta name="robots" content="noindex, nofollow">
+<meta name="robots" content="noindex, nofollow">
 @endif
 <meta property="og:title" content="{!! $collection->name !!}" />
 <meta property="og:description" content="{!! $collection->getShortTextAttribute($collection->text, 500) !!}" />
 <meta property="og:type" content="website" />
 <meta property="og:url" content="{!! Request::url() !!}" />
-<meta property="og:image" content="{!! asset_timed($collection->getHeaderImage()) !!}" />
+<meta property="og:image" content="{!! URL::to($collection->header_image_src) !!}" />
 <meta property="og:site_name" content="web umenia" />
 @foreach ($collection->getContentImages() as $image )
-    <meta property="og:image" content="{!! $image !!}" />
+<meta property="og:image" content="{!! $image !!}" />
 @endforeach
 @stop
 
@@ -32,51 +32,101 @@
 @section('content')
 
 @if ( ! $collection->hasTranslation(App::getLocale()) )
-    <section>
-        <div class="container top-section">
-            <div class="row">
-                @include('includes.message_untranslated')
+<section>
+    <div class="container top-section">
+        <div class="row">
+            @include('includes.message_untranslated')
+        </div>
+    </div>
+</section>
+@endif
+
+@component('components.header_carousel', ['item' => $collection]))
+    @slot('slideContent')
+        <h1>{!! $collection->name !!}</h1>
+        @if ($collection->type)
+        {{-- keep in one line to prevent formatting failures --}}
+        <h2><a href="{!! url_to( 'kolekcie', ['type' => $collection->type ]) !!}">{!! $collection->type !!}</a></h2>
+        @endif
+    @endslot
+@endcomponent
+
+<section class="collection content-header my-5">
+    <div class="collection-header">
+        <div class="container">
+            <div class="row text-center">
+                <div class="col-md-8 col-md-push-2">
+                    <div class="row">
+                        @if ($collection->items->count() != 0)
+                        <div class="col-sm-6">
+                            <div class="v-center min-h-3em">
+                                <a href="{!! url_to( 'kolekcie', ['author' => $collection->user->name ]) !!}">
+                                    {!! $collection->user->name !!}
+                                </a>
+                            </div>
+                        </div>
+                        <div class="col-sm-6">
+                            <div class="v-center min-h-3em">
+                                <i class='fa fa-calendar-o mr-3'></i>
+                                @date($collection->published_at)
+                            </div>
+                        </div>
+                        <div class="col-sm-6">
+                            <div class="v-center min-h-3em">
+                                <span>
+                                    {{trans('kolekcie.collections_items_count')}} <a
+                                       href="#artworks">{{trans_choice('general.artworks_counted', $collection->items->count(), ['artworks_count' => $collection->items->count()])}}</a>
+                                </span>
+                            </div>
+                        </div>
+                        <div class="col-sm-6">
+                            @if ($collection->reading_time)
+                            <div class="v-center min-h-3em">
+                                <span>
+                                    <i class='fa fa-clock-o mr-3'></i>
+                                    {!! $collection->reading_time !!}
+                                </span>
+                            </div>
+                            @endif
+                        </div>
+                        @else
+                        <div class="col-md-4 col-sm-6 col-xs-12">
+                            <div class="v-center min-h-3em">
+                                <a href="{!! url_to( 'kolekcie', ['author' => $collection->user->name ]) !!}">
+                                    {!! $collection->user->name !!}
+                                </a>
+                            </div>
+                        </div>
+                        <div class="col-md-4 col-sm-6 col-xs-12">
+                            <div class="v-center min-h-3em">
+                                <i class='fa fa-calendar-o mr-3'></i>
+                                @date($collection->published_at)
+                            </div>
+                        </div>
+                        <div class="col-md-4 col-md-push-0 col-sm-6 col-sm-push-6 col-xs-12">
+                            @if ($collection->reading_time)
+                            <div class="v-center min-h-3em">
+                                <span>
+                                    <i class='fa fa-clock-o mr-3'></i>
+                                    {!! $collection->reading_time !!}
+                                </span>
+                            </div>
+                            @endif
+                        </div>
+                        @endif
+                    </div>
+                </div>
             </div>
         </div>
-    </section>
-@endif
-
-<div class="webumeniaCarousel">
-
-@if ($collection->hasHeaderImage())
-<div class="header-image" style="background-image: url({!! $collection->getHeaderImage() !!}); color: {!! $collection->title_color !!}">
-@else
-<div class="header-image">
-@endif
-    <div class="outer-box">
-        <div class="inner-box">
-            <h1>{!! $collection->name !!}</h1>
-            <p class="bottom-space">
-                {{ trans_choice('general.artworks_counted', $collection->items()->count(), ['artworks_count' => $collection->items()->count()]) }} &nbsp;&middot;&nbsp;
-                {!! $collection->user->name !!} &nbsp;&middot;&nbsp;
-                {!! Carbon::parse($collection->published_at)->format('d. m. Y') !!}
-            </p>
-        </div>
     </div>
+</section>
 
-    <!-- share -->
-    {{-- <div class="shareon-container">
-        <div class="container text-right">
-            <div class="fb-like" data-href="http://dvekrajiny.sng.sk/" data-layout="button_count" data-action="like" data-show-faces="false" data-share="false"></div>
-            &nbsp;
-            <a href="https://twitter.com/share" class="twitter-share-button" data-count="true">Tweet</a>
-            <script>!function(d,s,id){var js,fjs=d.getElementsByTagName(s)[0],p=/^http:/.test(d.location)?'http':'https';if(!d.getElementById(id)){js=d.createElement(s);js.id=id;js.src=p+'://platform.twitter.com/widgets.js';fjs.parentNode.insertBefore(js,fjs);}}(document, 'script', 'twitter-wjs');</script>
-    </div>
-    </div> --}}
-</div>
-</div>
-
-<section class="collection content-section pb-0">
+<section class="collection my-5">
     <div class="collection-body">
         <div class="container">
             <div class="row">
-                <div class="col-md-8 col-md-offset-2 bottom-space description">
-                       {!! $collection->text !!}
+                <div class="col-md-8 col-md-offset-2 bottom-space description long-text">
+                    {!! $collection->text !!}
                 </div>
             </div>
         </div>
@@ -87,12 +137,12 @@
     @include('components.share_buttons', [
         'title' => $collection->name,
         'url' => $collection->getUrl(),
-        'img' => URL::to($collection->getHeaderImage()),
+        'img' => URL::to($collection->header_image_src),
         'class' => 'text-center mb-5'
     ])
 </section>
 
-<section class="collections content-section">
+<section class="collections">
     <div class="collections-body">
         <div class="container">
             <div class="row">
@@ -115,7 +165,6 @@
 @stop
 
 @section('javascript')
-{!! Html::script('js/slick.js') !!}
 {!! Html::script('js/components/share_buttons.js') !!}
 
 <script type="text/javascript">