+++ resources/views/layouts/admin.blade.php
@@ -25,7 +25,6 @@
         {!! Html::style('css/plugins/selectize.bootstrap3.css') !!}
         {!! Html::style('css/plugins/bootstrap-switch.css') !!}
         {!! Html::style('css/jquery-ui/jquery-ui.css') !!}
-        {!! Html::script('js/modernizr.custom.js') !!}
         {!! Html::style('/css/vue/vue-select.css') !!}
         {!! Html::style('/css/js/admin.css') !!}
 
@@ -202,8 +201,9 @@
     {!! Html::script('js/plugins/speakingurl.min.js') !!}
     {!! Html::script('js/plugins/bootstrap-switch.min.js') !!}
     {!! Html::script('js/jquery.collection.js') !!}
+    {!! Html::script('js/modernizr.custom.js') !!}
     {!! Html::script('js/selectize.min.js') !!}
-    
+
     <script type="text/javascript" src="{{ mix('/js/manifest.js') }}"></script>
 	<script type="text/javascript" src="{{ mix('/js/vendor.js') }}"></script>
 	<script type="text/javascript" src="{{ mix('/js/admin.js') }}"></script>
@@ -227,7 +227,7 @@
         var csrf = '{!!csrf_token()!!}';
         $( '.wysiwyg' ).ckeditor({
             language: 'sk',
-            extraAllowedContent: 'iframe[*]',
+            extraAllowedContent: 'iframe[*];*[data-*]{*}(*);*[class]{*}(*)',
             filebrowserUploadUrl: '/uploader?csrf_token='+csrf
         });