+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 
+## [2.22.0] - 2021-03-11
+### Added
+- add filtering to articles section
+
+### Changed
+- carousel & hero image refactoring, added img-srcset
+- unified main image storing for slides, collections and articles
+- new styling for artwork / article / collection, metadata blocks
+
 ## [2.21.2] - 2021-03-10
 ### Fixed
 - browsing between artworks within collection